It now works with low case too. +minor change
Major change: 
The change in cases does not affect the complexity and efficiency of the program much because of the implementation mean. 
Instead of adding 26 more characters to the selection pool of chars (this would make the program take much longer), I made the program turn the inputted word to lowcase regardless to work it through and then, after the letters match the target, check and change the cases of each position back. 
Minor change:
Reorganized and reuploaded the files in a new directory called "shakespeare".